startstopbutton/requirements.txt
redis
RPi.Gpio